+++ BackendServer/resources/env.example.json
@@ -6,6 +6,7 @@
   "DB_PASSWORD": "postgres",
   "DB_CONNECT_TIMEOUT_S": 0,
   "STEAM_WEB_API_KEY": "",
+  "STEAM_AUTH_WHITELIST": "",
   "Kestrel": {
     "Endpoints": {
       "Http1": {
