+++ Processors/Processors/PartyRpcProcessors.cs
@@ -718,321 +718,3 @@ public abstract class PartyRpcProcessorBase : WebsocketPacketProcessor
         }
     }
 }
-
-public class CreatePartyProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public CreatePartyProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.CreateV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        await RemovePlayerFromParties(ConnectionHandler.PlayerId, null);
-        Model.Party party =
-            await CreateParty(Packet.GetPayloadAsMessage<CreatePartyRequest>(), ConnectionHandler.PlayerId);
-        return await CreatePartyMessage(party);
-    }
-}
-
-public class JoinWithInviteCodeProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public JoinWithInviteCodeProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.JoinWithInviteCodeV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        string? inviteCode = ReadStringField(Packet.RequestPayload, "inviteCode", "invite_code", "code");
-        if (string.IsNullOrEmpty(inviteCode))
-        {
-            throw new InvalidDataException("JoinWithInviteCode request had no invite code");
-        }
-
-        Guid? partyId = await Model.Party.FindPartyIdByInviteCode(inviteCode);
-        if (partyId is null)
-        {
-            throw new InvalidDataException($"No party found for invite code {inviteCode}");
-        }
-
-        Model.Party party = await JoinPartyById(partyId.Value, ConnectionHandler.PlayerId, ReadJoinExt(Packet.RequestPayload));
-        return await CreatePartyMessageWithMemberFanout(party, ConnectionHandler.PlayerId);
-    }
-}
-
-public class JoinWithPartyIdProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public JoinWithPartyIdProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.JoinWithPartyIdV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        string? partyId = ReadStringField(Packet.RequestPayload, "partyId", "party_id");
-        if (string.IsNullOrEmpty(partyId))
-        {
-            throw new InvalidDataException("JoinWithPartyId request had no party id");
-        }
-
-        Model.Party party = await JoinPartyById(Guid.Parse(partyId), ConnectionHandler.PlayerId, ReadJoinExt(Packet.RequestPayload));
-        return await CreatePartyMessageWithMemberFanout(party, ConnectionHandler.PlayerId);
-    }
-}
-
-public class SendInviteProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public SendInviteProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.SendInviteV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        string? partyIdString = ReadStringField(Packet.RequestPayload, "partyId", "party_id");
-        string? inviteeString = ReadStringField(Packet.RequestPayload, "inviteePlayerId", "invitee_player_id", "playerId", "player_id");
-        if (string.IsNullOrEmpty(partyIdString) || string.IsNullOrEmpty(inviteeString))
-        {
-            throw new InvalidDataException("SendInvite request missing party id or invitee player id");
-        }
-
-        Guid partyId = Guid.Parse(partyIdString);
-        Guid inviteeId = Guid.Parse(inviteeString);
-        Model.Party party = await GetPartyOrThrow(partyIdString);
-        if (!Array.Exists(party.Members, m => m.PlayerId == ConnectionHandler.PlayerId))
-        {
-            throw new InvalidDataException($"Player {ConnectionHandler.PlayerId} is not in party {partyId}");
-        }
-
-        Guid inviteId = Guid.NewGuid();
-        string blob = "";
-        if (Packet.RequestPayload["extPartyInviteData"] is JsonObject extData)
-        {
-            blob = ReadStringField(extData, "blob") ?? "";
-        }
-
-        PendingInvites[inviteId] = new PendingPartyInvite(inviteId, partyId, ConnectionHandler.PlayerId, inviteeId, blob);
-        JsonObject response = new()
-        {
-            ["inviteId"] = inviteId.ToString()
-        };
-
-        JsonObject notification = new()
-        {
-            ["inviteId"] = inviteId.ToString(),
-            ["partyId"] = partyId.ToString(),
-            ["inviterInfo"] = await CreatePlayerInfoJson(ConnectionHandler.PlayerId),
-            ["extPartyInviteData"] = new JsonObject
-            {
-                ["blob"] = blob
-            }
-        };
-
-        return SpectreWebsocketMessage.From(response, postResponseActions:
-        [
-            async () => await SpectreWebsocket.SendNotificationToPlayerAsync(
-                inviteeId,
-                InviteReceivedNotificationType,
-                notification.ToJsonString())
-        ]);
-    }
-}
-
-public class RespondToInviteProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public RespondToInviteProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.RespondToInviteV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        string? inviteIdString = ReadStringField(Packet.RequestPayload, "inviteId", "invite_id");
-        if (string.IsNullOrEmpty(inviteIdString))
-        {
-            throw new InvalidDataException("RespondToInvite request had no invite id");
-        }
-
-        Guid inviteId = Guid.Parse(inviteIdString);
-        bool accepted = ReadBoolField(Packet.RequestPayload, false, "accepted", "accept", "bAccept");
-        if (!PendingInvites.TryRemove(inviteId, out PendingPartyInvite? invite))
-        {
-            throw new InvalidDataException($"No pending party invite found for id {inviteId}");
-        }
-        if (invite.InviteePlayerId != ConnectionHandler.PlayerId)
-        {
-            throw new InvalidDataException($"Player {ConnectionHandler.PlayerId} cannot respond to invite {inviteId}");
-        }
-
-        JsonObject responseNotification = new()
-        {
-            ["inviteId"] = inviteId.ToString(),
-            ["accepted"] = accepted,
-            ["inviteeInfo"] = await CreatePlayerInfoJson(ConnectionHandler.PlayerId)
-        };
-
-        if (!accepted)
-        {
-            return SpectreWebsocketMessage.From(new JsonObject(), postResponseActions:
-            [
-                async () => await SpectreWebsocket.SendNotificationToPlayerAsync(
-                    invite.InviterPlayerId,
-                    InviteResponseNotificationType,
-                    responseNotification.ToJsonString())
-            ]);
-        }
-
-        Model.Party party = await JoinPartyById(invite.PartyId, ConnectionHandler.PlayerId, ReadJoinExt(Packet.RequestPayload));
-        string partyJson = await BuildPartyJson(party);
-        return SpectreWebsocketMessage.From(partyJson, postResponseActions:
-        [
-            async () => await SpectreWebsocket.SendNotificationToPlayerAsync(
-                invite.InviterPlayerId,
-                InviteResponseNotificationType,
-                responseNotification.ToJsonString()),
-            async () => await NotifyOtherMembersBestEffort(party, ConnectionHandler.PlayerId, partyJson)
-        ]);
-    }
-}
-
-public class LeavePartyProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public LeavePartyProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.LeaveV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        await RemovePlayerFromParties(ConnectionHandler.PlayerId, null);
-        return SpectreWebsocketMessage.Empty();
-    }
-}
-
-public class UpdatePartyProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public UpdatePartyProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.UpdatePartyV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        UpdatePartyRequest request = Packet.GetPayloadAsMessage<UpdatePartyRequest>();
-        Model.Party party = await GetPartyOrThrow(request.PartyId);
-        ApplyPartyUpdate(party, request.RequestExt ?? new PartyUpdate());
-        await party.SyncToDatabase();
-        return await CreatePartyMessageWithMemberFanout(party, ConnectionHandler.PlayerId);
-    }
-}
-
-public class UpdatePartyPlayerProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public UpdatePartyPlayerProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.UpdatePartyPlayerV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        UpdatePartyPlayerRequest request = Packet.GetPayloadAsMessage<UpdatePartyPlayerRequest>();
-        Model.Party party = await GetPartyOrThrow(request.PartyId);
-        ApplyPartyPlayerUpdate(party, ConnectionHandler.PlayerId, request.RequestExt ?? new PartyPlayerUpdateData());
-        await party.SyncToDatabase();
-        return await CreatePartyMessageWithMemberFanout(party, ConnectionHandler.PlayerId);
-    }
-}
-
-public class SetReadyProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public SetReadyProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.SetReadyStateV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        SetReadyMessage request = Packet.GetPayloadAsMessage<SetReadyMessage>();
-        Model.Party party = await GetPartyOrThrow(request.PartyId);
-        ApplyReadyState(party, ConnectionHandler.PlayerId, request.Ready);
-        await party.SyncToDatabase();
-        return await CreatePartyMessageWithMemberFanout(party, ConnectionHandler.PlayerId);
-    }
-}
-
-public class EnterMatchmakingProcessor : PartyRpcProcessorBase, IWebsocketPacketProcessorSingleton
-{
-    [SetsRequiredMembers]
-    public EnterMatchmakingProcessor(SpectreRpcType rpcType) : base(rpcType)
-    {
-    }
-
-    public static SpectreRpcType GetRpcType()
-    {
-        return new SpectreRpcType("PartyRpc.EnterMatchmakingV1Request");
-    }
-
-    public override async Task<SpectreWebsocketMessage> ProcessPacket(SpectreWebsocketRequest Packet,
-        SpectreWebsocket ConnectionHandler)
-    {
-        EnterMatchmakingRequest request = Packet.GetPayloadAsMessage<EnterMatchmakingRequest>();
-        Model.Party party = await GetPartyOrThrow(request.PartyId);
-        WebsocketNotification[] postNotifs = await QueueMatchmakingNotifications(party, ConnectionHandler);
-        return await CreatePartyMessage(party, postNotifs);
-    }
-}
